+++ common/src/main/java/com/frostwire/search/archiveorg/ArchiveorgTorrentSearchResult.java
@@ -28,9 +28,9 @@ import com.frostwire.search.torrent.TorrentSearchResult;
  */
 public class ArchiveorgTorrentSearchResult extends ArchiveorgCrawledSearchResult implements TorrentSearchResult {
 
-    private final long size;
+    private final double size;
 
-    public ArchiveorgTorrentSearchResult(ArchiveorgSearchResult sr, ArchiveorgFile file, long size) {
+    public ArchiveorgTorrentSearchResult(ArchiveorgSearchResult sr, ArchiveorgFile file, double size) {
         super(sr, file);
         this.size = size;
     }
@@ -56,7 +56,7 @@ public class ArchiveorgTorrentSearchResult extends ArchiveorgCrawledSearchResult
     }
 
     @Override
-    public long getSize() {
+    public double getSize() {
         return size;
     }
 }
