+++ desktop/src/com/frostwire/gui/bittorrent/BTDownloadMediator.java
@@ -962,7 +962,7 @@ public final class BTDownloadMediator extends AbstractTableMediator<BTDownloadRo
         });
     }
 
-    public void openHttp(final String httpUrl, final String title, final String saveFileAs, final long fileSize) {
+    public void openHttp(final String httpUrl, final String title, final String saveFileAs, final double fileSize) {
         GUIMediator.safeInvokeLater(() -> {
             final HttpDownload downloader = new HttpDownload(httpUrl, title, saveFileAs, fileSize, null, false, true) {
                 @Override
